+++ tests/testthat/helpers.R
@@ -37,12 +37,22 @@ test_knit_report <- function(report_type, outfile_ext){
                       report_type = report_type,
                       interactive = FALSE
       )
-      expect_no_error(
-        rmarkdown::render(file.path(report_name, paste0(report_name, '.Rmd')),
-                          output_format = output_format,
-                          quiet = TRUE,
-                          clean = FALSE)
+      # test knit in a separate R session
+      rs <- callr::r_session$new()
+      on.exit(rs$close(), add = TRUE)
+      res <- rs$run_with_output(
+        function(...) rmarkdown::render(envir = globalenv(), ...),
+        args = list(
+          file.path(report_name, paste0(report_name, '.Rmd')),
+          output_format = output_format,
+          quiet = TRUE,
+          clean = FALSE
+        )
       )
+      # propagate full error stack from subprocess into main process
+      if (!is.null(res$error)) {
+        stop(paste(capture.output(res$error), collapse = '\n'))
+      }
     })
     outfile_sans_ext <- file.path(temp_dir, report_name, report_name)
     expect_true(