+++ NEWS.md
@@ -11,6 +11,7 @@ Improvements for users
 Improvements for package contributors and maintainers
 * Repair unit tests to work with `testthat 3.3.0` (#317)
 * Remove empty "examples" in template study schema to avoid minor issues (#319)
+* Run test knits in a `callr` session to isolate Rmd library calls from main R process (#320)
 
 # VISCtemplates 2.0.2